First go at management helper class, rabbitmq_vhost, rabbitmq_user
Everything will be driven by a client class variable that is just a connection
to the management API, effectively. It provides some nice methods to easily
accomplish our task, and eliminates almost all of the state-checking that might
be necessary when shelling out.

There is some verification to do, as well as some fixes for
updated_by_last_action. It's also not entirely clear how hard we should work to
verify our changes did or did not actually commit. Probably should do something
though.