Remove all defaults for configuration.
RabbitMQ will automatically use default values if it needs to. We shouldn't do
this ourselves because in doing so, we force people to re-passin the values of
all configuration each time. For now, that just means the environment and
rabbitmq.config, but it's still an annoyance. This way, if you provide a nil
value for a parameter, it's okay, while if you pass in an empty hash, it will
blank out the file. This is more consistent behavior than before, where it
would assign values for you from under your feet.