Take two of management module
This attempts to just create one client connection ever. It is then up to the
caller to make sure they set all the proper attributes before the first
rabbitmq resource they call (besides the default or _config resources).

A future update would be to verify the options of the client connection are
consistent with what's in attributes, but this is an acceptable pass, and
eliminates much of the garbage from before. I've also verified that the right
values get updated if you set attributes.