Revert permissions for rabbitmq_binding after creation.
We have to elevate the admin user's access to the virtualhost upon which a
RabbitMQ binding is created, as read, write, and configure are all required to
declare a binding.

However, this is kind of a strange case to uniquely ruin the semi-secure model
of configure-only for, so we can set the permissions back to the way they were
after we declare the binding.