Finally render the Erlang cookie correctly.
The steps for changing the Erlang cookie on an already-running RabbitMQ node
are to stop the entire Erlang node, change the cookie, start the Erlang node
(and RabbitMQ by extension), and then reset RabbitMQ data using rabbitmqctl.

This also makes sure to check if the cookie has changed, lest we stop/start
RabbitMQ every time.