I really hate Ruby/Chef sometimes.
Chef/Ruby does some weird coercion under the covers which means that by the
time node[:rabbitmq] gets to this library, the keys are strings instead of
symbols (even if you pass them in as such). This means there was both a for
example :admin_port and "admin_port" key to the hash, and they weren't getting
merged. We're gonna use 1.8 hash syntax because it works that way. PRs
accepted.

Please eliminate Chef from my life.